+++ services/galley/src/Galley/API/Federation.hs
@@ -17,9 +17,6 @@
 module Galley.API.Federation where
 
 import Data.Containers.ListUtils (nubOrd)
-import Data.Domain (Domain)
-import Data.Id (UserId)
-import qualified Data.Map as Map
 import Data.Qualified (Qualified (..))
 import qualified Galley.API.Mapping as Mapping
 import Galley.API.Util (fromRegisterConversation, pushConversationEvent, viewFederationDomain)
@@ -29,7 +26,7 @@ import Imports
 import Servant (ServerT)
 import Servant.API.Generic (ToServantApi)
 import Servant.Server.Generic (genericServerT)
-import Wire.API.Conversation.Member (Member, memId)
+import Wire.API.Conversation.Member (OtherMember (..), memId)
 import Wire.API.Event.Conversation
 import Wire.API.Federation.API.Galley (ConversationMemberUpdate (..), GetConversationsRequest (..), GetConversationsResponse (..), RegisterConversation (..))
 import qualified Wire.API.Federation.API.Galley as FederationAPIGalley
@@ -46,12 +43,14 @@ federationSitemap =
 registerConversation :: RegisterConversation -> Galley ()
 registerConversation rc = do
   localDomain <- viewFederationDomain
-  let localUsers = fmap (toQualified localDomain) . getLocals $ localDomain
-      localUserIds = map qUnqualified localUsers
+  let localUsers =
+        foldMap (\om -> guard (qDomain (omQualifiedId om) == localDomain) $> omQualifiedId om)
+          . rcMembers
+          $ rc
+      localUserIds = fmap qUnqualified localUsers
   unless (null localUsers) $ do
     Data.addLocalMembersToRemoteConv localUserIds (rcCnvId rc)
-  forM_ localUsers $ \usr -> do
-    c <- fromRegisterConversation usr rc
+  forM_ (fromRegisterConversation localDomain rc) $ \(mem, c) -> do
     let event =
           Event
             ConvCreate
@@ -59,12 +58,7 @@ registerConversation rc = do
             (rcOrigUserId rc)
             (rcTime rc)
             (EdConversation c)
-    pushConversationEvent event [qUnqualified usr] []
-  where
-    getLocals :: Domain -> [Member]
-    getLocals localDomain = fromMaybe [] . Map.lookup localDomain . rcMembers $ rc
-    toQualified :: Domain -> Member -> Qualified UserId
-    toQualified domain mem = Qualified (memId mem) domain
+    pushConversationEvent event [memId mem] []
 
 getConversations :: GetConversationsRequest -> Galley GetConversationsResponse
 getConversations (GetConversationsRequest qUid gcrConvIds) = do