+++ CHANGELOG.md
@@ -16,9 +16,11 @@
 
 # Unreleased
 
-## Internal changes
+# Internal changes
 
 * schema-profunctor: add `optField` combinator and corresponding documentation (#1621, #1624).
+* [Federation] Let a receiving backend decide conversation attribute specifics of its users
+  added to a new conversation via `POST /federation/register-conversation` (#1622)
 
 ## Documentation